requirements.txt
flask==3.1.0
flask_sqlalchemy==3.1.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
python-telegram-bot==20.7
python-dotenv
alembic
